Move messaging functions to a separate file/oject
