+++ common.js
@@ -1,125 +1,12 @@
 const path = require("path");
 const {platform} = require("rise-common-electron");
-const heartbeat = require("./heartbeat");
 const portedPlatform = require("./platform");
-const EventEmitter = require('events');
 const HttpProxyAgent = require("proxy-agent");
 const HttpsProxyAgent = require("https-proxy-agent");
 global.log = global.log || {error:console.log,debug:console.log};
-let lmsClient = null, ipcConnection = null;
-const ipc = require('node-ipc');
 
 function getDisplaySettingsFileName() {
   return path.join(getInstallDir(), "RiseDisplayNetworkII.ini");
-}
-
-function connect(id) {
-  if (ipcConnection) {
-    return ipcConnection;
-  }
-  ipcConnection = new Promise((resolve)=>{
-    if (lmsClient) {
-      resolve(lmsClient);
-    } else {
-      ipc.config.id   = id;
-      ipc.config.retry= 1500;
-      ipc.config.silent = !process.argv.includes("--debug");
-
-      ipc.connectTo(
-          'lms',
-          () => {
-              ipc.of.lms.on(
-                  'connect',
-                  () => {
-                      ipc.log('## connected to lms ##', ipc.config.id);
-                      ipc.of.lms.emit("connected", {id: ipc.config.id, client: ipc.config.id});
-
-                      lmsClient = {
-                        broadcastMessage: (message) => {
-                          ipc.of.lms.emit('message', message);
-                        },
-                        getClientList: () => {
-                          ipc.of.lms.emit("clientlist-request");
-                        },
-                        toMessagingService: (message) => {
-                          ipc.of.lms.emit("message", Object.assign({}, message, {through: "ms"}));
-                        },
-                        toLocalWS: (message) => {
-                          ipc.of.lms.emit("message", Object.assign({}, message, {through: "ws"}));
-                        },
-                        receiveMessages: () => {
-                          let receiver = new EventEmitter();
-                          ipc.of.lms.on(
-                              'message',
-                              function(message){
-                                  ipc.log('got a message from lms : ', message);
-                                  receiver.emit("message",message);
-                              }
-                          );
-                          return receiver;
-                        }
-                      }
-                      resolve(lmsClient);
-                  }
-              );
-              ipc.of.lms.on(
-                  'disconnect',
-                  function(){
-                      ipc.log('disconnected from lms');
-                  }
-              );
-
-              ipc.of.lms.on(
-                  'error',
-                  function(){
-                      ipc.log('error from lms');
-                  }
-              );
-          }
-      );
-    }
-  });
-  return ipcConnection;
-}
-
-function disconnect() {
-  ipc.disconnect('lms');
-  ipcConnection = null;
-  lmsClient = null;
-}
-
-function broadcastMessage(message) {
-  connect(message.from).then((client)=>{
-    client.broadcastMessage(message);
-  });
-}
-
-function broadcastToLocalWS(message) {
-  connect(message.from).then((client)=>{
-    client.toLocalWS(message);
-  });
-}
-
-function getClientList(id) {
-  connect(id).then((client)=>{
-    client.getClientList();
-  });
-}
-
-function receiveMessages(id) {
-  return new Promise((resolve)=>{
-    connect(id).then((client)=>{
-      resolve(client.receiveMessages());
-
-      heartbeat.startHearbeatInterval(id);
-    });
-  });
-}
-
-function sendToMessagingService(message) {
-  connect(message.from).then((client)=>{
-    client.toMessagingService(message);
-  });
 }
 
 function getDisplaySettings() {
@@ -284,13 +171,6 @@ module.exports = {
     }
   },
   getLatestVersionInManifest,
-  connect,
-  disconnect,
-  broadcastMessage,
-  broadcastToLocalWS,
-  getClientList,
-  sendToMessagingService,
-  receiveMessages,
   getProxyAgents,
   moduleIsBackgroundTask(name) {
     return module.exports.getModulePackage(name).backgroundTask;
